Change fadeMiddle to be an int instead of float for moral reasons

+++ FloodForge/assets/shaders/connection.frag
@@ -6,7 +6,7 @@ out vec4 color;
 uniform vec4 tintColor;
 uniform vec4 tintColorB;
 uniform float widthClip;
-uniform float fadeMiddle;
+uniform int fadeMiddle;
 
 vec3 lerp(vec3 a, vec3 b, float t) {
 	return (b - a) * t + a;
